# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PetStore API",
        "version": "1.0"
    },
    "paths": {
        "/api/pets": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Get list of pets",
                "operationId": "7b44ba1ad8a38f7865a2845fa9ebbfe3",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pet"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pets"
                ],
                "summary": "Add a new pet to the store",
                "operationId": "9523721805529edb47b46332d56f2826",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "category": {
                                        "properties": {
                                            "id": {
                                                "description": "ID of the category",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Name of the category",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "name": {
                                        "description": "Name of the pet",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status of the pet",
                                        "type": "string"
                                    },
                                    "photoUrls": {
                                        "description": "List of photo URLs",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "tags": {
                                        "description": "List of tags",
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Tag"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pet created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pet"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pets/{id}": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Find pet by ID",
                "operationId": "25e25b195a59c4b940a33f9961ad6eed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the pet to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pets"
                ],
                "summary": "Update an existing pet",
                "operationId": "652be7eab297dff87f093d849bdecb12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the pet to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category": {
                                        "properties": {
                                            "id": {
                                                "description": "ID of the category",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Name of the category",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "name": {
                                        "description": "Name of the pet",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status of the pet",
                                        "type": "string"
                                    },
                                    "photoUrls": {
                                        "description": "List of photo URLs",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "tags": {
                                        "description": "List of tags",
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Tag"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pet updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pets"
                ],
                "summary": "Deletes a pet",
                "operationId": "dab25348cae9362e48ade0e7cf34c465",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the pet to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pet deleted"
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                }
            }
        },
        "/api/pets/status": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Find pets by status",
                "operationId": "7a3ea3e52e2ab4f3f84808c2b25ce1a2",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status to filter pets",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pet"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pets/tags": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Find pets by tags",
                "operationId": "8748201d0ab2db684d45baddd44b634d",
                "parameters": [
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Comma-separated list of tag names to filter pets",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pet"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pets/{petId}/uploadImage": {
            "post": {
                "tags": [
                    "Pets"
                ],
                "summary": "Uploads an image",
                "operationId": "cbcc30db15a17df9d80db02984e90825",
                "parameters": [
                    {
                        "name": "petId",
                        "in": "path",
                        "description": "ID of the pet to upload image",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Image uploaded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "description": "URL of the uploaded image",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Image upload failed"
                    }
                }
            }
        },
        "/api/pets/{petId}": {
            "post": {
                "tags": [
                    "Pets"
                ],
                "summary": "Updates a pet in the store with form data",
                "operationId": "adf7e6d224a88a73f5a17f25abb2f623",
                "parameters": [
                    {
                        "name": "petId",
                        "in": "path",
                        "description": "ID of the pet to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the pet",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status of the pet",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pet updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Pet": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Pets",
            "description": "Pets"
        }
    ]
}